+++ services/workflows-service/src/storage/storage.controller.internal.ts
@@ -8,24 +8,12 @@ import type { Response } from 'express';
 import { StorageService } from './storage.service';
 import * as errors from '../errors';
 import { fileFilter } from './file-filter';
-import {
-  createPresignedUrlWithClient,
-  downloadFileFromS3,
-  manageFileByProvider,
-} from '@/storage/get-file-storage-manager';
-import { AwsS3FileConfig } from '@/providers/file/file-provider/aws-s3-file.config';
-import path from 'path';
-import os from 'os';
-import { File } from '@prisma/client';
-import { z } from 'zod';
-import { HttpFileService } from '@/providers/file/file-provider/http-file.service';
+import { manageFileByProvider } from '@/storage/get-file-storage-manager';
 import { ProjectIds } from '@/common/decorators/project-ids.decorator';
 import type { TProjectId, TProjectIds } from '@/types';
 import { CurrentProject } from '@/common/decorators/current-project.decorator';
-import { isBase64 } from '@/common/utils/is-base64/is-base64';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { HttpService } from '@nestjs/axios';
-import mime from 'mime';
 import { getFileMetadata } from '@/common/get-file-metadata/get-file-metadata';
 
 // Temporarily identical to StorageControllerExternal
@@ -114,67 +102,18 @@ export class StorageControllerInternal {
     @Res() res: Response,
     @Query('format') format: string,
   ) {
-    // currently ignoring user id due to no user info
-    const persistedFile = await this.service.getFileById({ id }, projectIds, {});
+    const { mimeType, signedUrl, filePath } = await this.service.fetchFileContent({
+      id,
+      projectIds,
+      format,
+    });
 
-    if (!persistedFile) {
-      throw new errors.NotFoundException('file not found');
-    }
-
-    const mimeType =
-      persistedFile.mimeType ||
-      mime.getType(persistedFile.fileName || persistedFile.uri || '') ||
-      undefined;
-
-    if (persistedFile.fileNameInBucket && format === 'signed-url') {
-      const signedUrl = await createPresignedUrlWithClient({
-        bucketName: AwsS3FileConfig.getBucketName(process.env) as string,
-        fileNameInBucket: persistedFile.fileNameInBucket,
-        mimeType,
-      });
-
+    if (signedUrl) {
       return res.json({ signedUrl, mimeType });
     }
 
     res.set('Content-Type', mimeType || 'application/octet-stream');
 
-    if (persistedFile.fileNameInBucket) {
-      const localFilePath = await downloadFileFromS3(
-        AwsS3FileConfig.getBucketName(process.env) as string,
-        persistedFile.fileNameInBucket,
-      );
-
-      return res.sendFile(localFilePath, { root: '/' });
-    }
-
-    if (!isBase64(persistedFile.uri) && this._isUri(persistedFile)) {
-      const downloadFilePath = await this.__downloadFileFromRemote(persistedFile);
-
-      return res.sendFile(this.__getAbsoluteFilePAth(downloadFilePath));
-    }
-
-    return res.sendFile(this.__getAbsoluteFilePAth(persistedFile.fileNameOnDisk));
-  }
-
-  private __getAbsoluteFilePAth(filePath: string) {
-    if (!path.isAbsolute(filePath)) return filePath;
-
-    const rootDir = path.parse(os.homedir()).root;
-
-    return path.join(rootDir, filePath);
-  }
-
-  private async __downloadFileFromRemote(persistedFile: File) {
-    const localeFilePath = `${os.tmpdir()}/${persistedFile.id}`;
-    const downloadedFilePath = await new HttpFileService({
-      client: this.httpService,
-      logger: this.logger,
-    }).download(persistedFile.uri, localeFilePath);
-
-    return downloadedFilePath;
-  }
-
-  _isUri(persistedFile: File) {
-    return z.string().url().safeParse(persistedFile.uri).success;
+    return res.sendFile(filePath!);
   }
 }