+++ packages/common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/common
 
+## 0.9.36
+
+### Patch Changes
+
+- version bump
+
+## 0.9.35
+
+### Patch Changes
+
+- bump
+
+## 0.9.34
+
+### Patch Changes
+
+- Bump
+
+## 0.9.33
+
+### Patch Changes
+
+- Bump
+
 ## 0.9.32
 
 ### Patch Changes