+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/config
 
+## 1.1.18
+
+### Patch Changes
+
+- Bump
+
+## 1.1.17
+
+### Patch Changes
+
+- Bump
+
 ## 1.1.16
 
 ### Patch Changes