+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/eslint-config-react
 
+## 2.0.18
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.18
+
+## 2.0.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.17
+
 ## 2.0.16
 
 ### Patch Changes