+++ packages/rules-engine/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/rules-engine-lib
 
+## 0.5.18
+
+### Patch Changes
+
+- Bump
+
+## 0.5.17
+
+### Patch Changes
+
+- Bump
+
 ## 0.5.16
 
 ### Patch Changes