+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/websocket-service
 
+## 0.1.18
+
+### Patch Changes
+
+- Bump
+
+## 0.1.17
+
+### Patch Changes
+
+- Bump
+
 ## 0.1.16
 
 ### Patch Changes