+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @ballerine/workflows-service
 
+## 0.7.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.36
+  - @ballerine/workflow-core@0.6.48
+  - @ballerine/workflow-node-sdk@0.6.48
+
+## 0.7.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.35
+  - @ballerine/workflow-core@0.6.47
+  - @ballerine/workflow-node-sdk@0.6.47
+
+## 0.7.49
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.34
+  - @ballerine/workflow-core@0.6.46
+  - @ballerine/workflow-node-sdk@0.6.46
+
+## 0.7.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.45
+  - @ballerine/workflow-node-sdk@0.6.45
+
+## 0.7.47
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.33
+  - @ballerine/workflow-core@0.6.44
+  - @ballerine/workflow-node-sdk@0.6.44
+
 ## 0.7.46
 
 ### Patch Changes